撤回消息

+++ fim_server/models/file_models/enetr.go
@@ -10,7 +10,7 @@ type FileModel struct {
 	Uid    uuid.UUID `json:"uid"`     // 文件唯一ID  /api/file/{uid}
 	UserId uint      `json:"user_id"` // 用户id
 	Name   string    `json:"name"`    // 文件名称
-	Size   float64   `json:"size"`    // 文件大小
+	Size   int64     `json:"size"`    // 文件大小
 	Path   string    `json:"path"`    // 文件路径
 	Hash   string    `json:"hash"`    // 文件哈希
 }
